# Avoid overhead of comparision function pointer calls in lpFind()