@sundb 2024-08-29
```
taskset -c 10-15 ./src/redis-server --save ""
```

```
taskset -c 16-20 memtier_benchmark --pipeline 30 --test-time 60 -c 1000 -t 2 -x 3 --hide-histogram --command='HSET h a 1 b 2 c 3 d 4 e 5 f 6 g 7 h 8 i 9 j 10 k 11 l 12 m 13 n 14 o 15 p 16 q 17 r 18 s 19 t 20'
```

before #13279 (95cbe879e54ab19d0120eac9fa10240759a129e1):
```
AGGREGATED AVERAGE RESULTS (3 runs)
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Hsets      413235.66       144.76332       150.52700       210.94300       230.39900    127925.49 
Totals     413235.66       144.76332       150.52700       210.94300       230.39900    255850.99
```

#13279 (a25b15392c3f50909ab2dafbcbdb3794ada29620):
```
AGGREGATED AVERAGE RESULTS (3 runs)
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Hsets      383770.26       156.34110       161.79100       223.23100       240.63900    118803.88 
Totals     383770.26       156.34110       161.79100       223.23100       240.63900    237607.76
```

this patch with #13279: (still other regissions after #13279)
```
AGGREGATED AVERAGE RESULTS (3 runs)
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Hsets      409445.97       146.62460       152.57500       212.99100       232.44700    126752.32 
Totals     409445.97       146.62460       152.57500       212.99100       232.44700    253504.64
```