@fcostaoliveira 2024-08-29
### CE Performance Automation : step 1 of 2 (build) DONE.

This comment was automatically generated given a benchmark was triggered.
Started building at 2024-08-29 14:49:10.655918 and took 86 seconds.
You can check each build/benchmark progress in grafana:
   - git hash: 600a0b4220dcfce8c86e0deec6777c46e42ff3a1
   - git branch: sundb:regression_lpFind
   - commit date and time: n/a
   - commit summary: n/a
   - test filters:
       - command priority lower limit: 0
       - command priority upper limit: 10000
       - test name regex: .*
       - command group regex: .*

You can check a comparison in detail via the [grafana link](https://benchmarksredisio.grafana.net/d/edsxdsrbexhc0f/ce-benchmark-run-status?orgId=1&var-benchmark_work_stream=1724943039109-0)