@fcostaoliveira 2024-08-30
### CE Performance Automation : step 2 of 2 (benchmark) RUNNING...

This comment was automatically generated given a benchmark was triggered.

Started benchmark suite at 2024-08-30 16:57:57.218523 and took 0 seconds up until now.
Status: [--------------------------------------------------------------------------------] 0.0% completed.

In total will run 107 benchmarks.
    - 107 pending.
    - 0 completed:
      - 0 successful.
      - 0 failed.
You can check a the status in detail via the [grafana link](https://benchmarksredisio.grafana.net/d/edsxdsrbexhc0f/ce-benchmark-run-status?orgId=1&var-benchmark_work_stream=1724951472380-0)